agent_sdk/requirements.txt
openai-agents>=0.14.0
mysql-connector-python>=9.0.0
pandas
python-dotenv>=1.0.0
fastapi>=0.115.0
uvicorn>=0.30.0
pydantic>=2.0.0
